requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-ollama
python-dotenv